+++ packages/react-router/lib/context.ts
@@ -1,18 +1,18 @@
 import * as React from "react";
 import type {
-  AgnosticRouteMatch,
   AgnosticIndexRouteObject,
   AgnosticNonIndexRouteObject,
+  AgnosticRouteMatch,
   History,
+  LazyRouteFunction,
   Location,
+  Action as NavigationType,
   RelativeRoutingType,
   Router,
   StaticHandlerContext,
   To,
   TrackedPromise,
-  LazyRouteFunction,
 } from "@remix-run/router";
-import type { Action as NavigationType } from "@remix-run/router";
 
 // Create react-specific types from the agnostic types in @remix-run/router to
 // export from react-router
@@ -100,7 +100,7 @@ export interface NavigateOptions {
  * A Navigator is a "location changer"; it's how you get to different locations.
  *
  * Every history instance conforms to the Navigator interface, but the
- * distinction is useful primarily when it comes to the low-level <Router> API
+ * distinction is useful primarily when it comes to the low-level `<Router>` API
  * where both the location and a navigator must be provided separately in order
  * to avoid "tearing" that may occur in a suspense-enabled app if the action
  * and/or location were to be read directly from the history instance.